test: consolidate filemonitor integration tests, drop redundant qhy-focuser test (#162)
Phase 3 of docs/plans/build-optimization-test-reorganization.md.

filemonitor: merge test_server.rs, test_reload.rs, and test_cli.rs into a
single tests/test_integration.rs. Drops test_cli_invalid_config — already
covered by the BDD scenario "Reject invalid configuration sources" in
configuration.feature, which exercises both an invalid JSON path and a
nonexistent path. 3 integration binaries become 1.

qhy-focuser: remove test_server_returns_configured_focuser_name from
tests/test_lib.rs — already covered by the BDD scenario "Device reports
configured name" in device_metadata.feature, which spawns the real
qhy-focuser binary and queries /api/v1/focuser/0/name through an Alpaca
client.

docs/skills/testing.md §6.4: update Test File Organization to reflect the
consolidated layout.
